+++ QuizWebsite/src/main/java/org/ja/dao/MatchesDao.java
@@ -2,12 +2,23 @@ package org.ja.dao;
 
 import org.apache.commons.dbcp2.BasicDataSource;
 import org.ja.model.OtherObjects.Match;
+
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
+/*
+create table matches(
+    match_id bigint primary key auto_increment,
+    question_id bigint not null,
+    left_match text not null,
+    right_match text not null,
 
+    foreign key (question_id) references questions(question_id) on delete cascade
+);
+
+ */
 public class MatchesDao {
     private final BasicDataSource dataSource;
     private long cnt=0;
@@ -16,7 +27,10 @@ public class MatchesDao {
     }
 
     public void insertMatch(Match match) {
-        String sql = "INSERT INTO matches (question_id, left_match, right_match) VALUES (?,?,?)";
+//        if(contains(match)) {
+//            return;
+//        }
+        String sql = "insert into matches (question_id, left_match, right_match) values(?,?,?)";
         try(Connection c = dataSource.getConnection();
             PreparedStatement ps = c.prepareStatement(sql, PreparedStatement.RETURN_GENERATED_KEYS)) {
 
@@ -28,7 +42,7 @@ public class MatchesDao {
             try (ResultSet rs = ps.getGeneratedKeys()){
                 if(rs.next()) {
                     cnt++;
-                    match.setMatchId(rs.getLong(1));
+                    match.setMatchId(rs.getLong("match_id"));
                 }
             }
 
@@ -39,14 +53,17 @@ public class MatchesDao {
     }
 
     public void removeMatch(long matchId) {
+        if(getMatchById(matchId) == null) {
+            return;
+        }
         String sql = "delete from matches where match_id = ?";
         try (Connection c = dataSource.getConnection();
             PreparedStatement ps = c.prepareStatement(sql)){
 
             ps.setLong(1, matchId);
 
-            if(ps.executeUpdate() > 0)
-                cnt--;
+            ps.executeUpdate();
+            cnt--;
         } catch(SQLException e){
             throw new RuntimeException("Error removing Match from database", e);
         }
@@ -71,7 +88,6 @@ public class MatchesDao {
 
         return matches;
     }
-
     public boolean contains(Match match) {
         String sql = "SELECT COUNT(*) FROM matches WHERE match_id = ? AND question_id=?" +
                 "AND left_match=? AND right_match = ?";
@@ -94,7 +110,6 @@ public class MatchesDao {
             throw new RuntimeException("Error checking user existence", e);
         }
     }
-
     public Match getMatchById(long matchId) {
         String sql = "select * from matches where match_id = "+matchId;
         try(Connection c=dataSource.getConnection(); PreparedStatement ps=c.prepareStatement(sql)){
@@ -107,11 +122,9 @@ public class MatchesDao {
             throw new RuntimeException(e);
         }
     }
-
     public long getCount(){
         return cnt;
     }
-
     private Match retrieveMatch(ResultSet rs) throws SQLException {
         return new Match(rs.getLong("match_id"), rs.getLong("question_id"),
                 rs.getString("left_match"), rs.getString("right_match"));
